Implemented MLS methods pop_multilangstring and fixes on other MLS methods.

+++ langstring/multilangstring.py
@@ -482,6 +482,7 @@ class MultiLangString:
 
     # ----- GET METHODS -----
 
+    @Validator.validate_simple_type
     def get_langs(self, casefold: bool = False) -> list[str]:
         """Return a list with all languages in the MultiLangString."""
         return [lang.lower() for lang in self.mls_dict.keys()] if casefold else list(self.mls_dict.keys())
@@ -497,15 +498,21 @@ class MultiLangString:
             raise TypeError(f"Invalid argument 'text' received. Expected 'str', got '{type(text).__name__}'.")
         if not isinstance(lang, str):
             raise TypeError(f"Invalid argument 'lang' received. Expected 'str', got '{type(lang).__name__}'.")
+
         return LangString(text=text, lang=lang) if self.contains_entry(text=text, lang=lang) else default
 
     def get_setlangstring(self, lang: str, default: Optional[Any] = None) -> Union[SetLangString, Any]:
         if not isinstance(lang, str):
             raise TypeError(f"Invalid argument 'lang' received. Expected 'str', got '{type(lang).__name__}'.")
+
         return SetLangString(texts=self.mls_dict[lang], lang=lang) if self.contains_lang(lang=lang) else default
 
     def get_multilangstring(self, langs: list[str], default: Optional[Any] = None) -> Union["MultiLangString", Any]:
-        # TODO: Add type validation
+        if not isinstance(langs, list):
+            raise TypeError(f"Invalid argument 'langs' received. Expected 'list', got '{type(langs).__name__}'.")
+        if not all(isinstance(item, str) for item in langs):
+            raise TypeError("Invalid argument 'langs' received. Not all elements in the list are strings.")
+
         if all(lang in self.mls_dict for lang in langs):
             return MultiLangString(mls_dict={lang: self.mls_dict[lang] for lang in langs if lang in self.mls_dict},
                                    pref_lang=self.pref_lang)
@@ -515,7 +522,11 @@ class MultiLangString:
     # ----- POP METHODS -----
 
     def pop_langstring(self, text: str, lang: str, default: Optional[Any] = None) -> Union[LangString, Any]:
-        # TODO: Add type validation
+        if not isinstance(text, str):
+            raise TypeError(f"Invalid argument 'text' received. Expected 'str', got '{type(text).__name__}'.")
+        if not isinstance(lang, str):
+            raise TypeError(f"Invalid argument 'lang' received. Expected 'str', got '{type(lang).__name__}'.")
+
         if self.contains_entry(text=text, lang=lang):
             new_ls = self.get_langstring(text=text, lang=lang)
             self.remove_entry(text=text, lang=lang)
@@ -523,7 +534,9 @@ class MultiLangString:
         return default
 
     def pop_setlangstring(self, lang: str, default: Optional[Any] = None) -> Union[SetLangString, Any]:
-        # TODO: Add type validation
+        if not isinstance(lang, str):
+            raise TypeError(f"Invalid argument 'lang' received. Expected 'str', got '{type(lang).__name__}'.")
+
         if self.contains_lang(lang=lang):
             new_sls = self.get_setlangstring(lang=lang)
             self.remove_lang(lang=lang)
@@ -531,8 +544,17 @@ class MultiLangString:
         return default
 
     def pop_multilangstring(self, langs: list[str], default: Optional[Any] = None) -> Union["MultiLangString", Any]:
-        # TODO: TO BE IMPLEMENTED
-        pass
+        if not isinstance(langs, list):
+            raise TypeError(f"Invalid argument 'langs' received. Expected 'list', got '{type(langs).__name__}'.")
+        if not all(isinstance(item, str) for item in langs):
+            raise TypeError("Invalid argument 'langs' received. Not all elements in the list are strings.")
+
+        new_mls = self.get_multilangstring(langs, None)
+        if new_mls:
+            for lang in langs:
+                self.remove_lang(lang)
+            return new_mls
+        return default
 
     # ----- GENERAL METHODS -----
